Add Firebase authentication Logic
