+++ form/package.json
@@ -8,6 +8,7 @@
     "npm": "^7.11.1",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-firebase-hooks": "^3.0.4",
     "react-scripts": "4.0.1"
   },
   "scripts": {
